add item filter in fetch and base_id in base repo

+++ src/repository/adapters/base_sql_repo.py
@@ -8,39 +8,48 @@ import inspect
 T = TypeVar("T")
 
 
+def clearNone(original: dict):
+    filtered = {k: v for k, v in original.items() if v is not None}
+    original.clear()
+    original.update(filtered)
+    return filtered
+
+
 class BaseRepo(BaseRepository[T]):
     def __init__(
         self,
-        session_factory: Callable[
-            ..., AbstractContextManager[Session]
-        ],
+        session_factory: Callable[..., AbstractContextManager[Session]],
         model,
         primary_key_identifier="id",
     ) -> None:
         self.session_factory = session_factory
         self.model = model
         self.identifier = primary_key_identifier
+        self.auto_model_conversion = False
 
     def add(self, **schema):
         query = self.model(**schema)
+        # if self.auto_model_conversion:
+        #     query = schema
         with self.session_factory() as session:
             try:
                 session.add(query)
                 session.commit()
                 session.refresh(query)
             except IntegrityError as error:
-                raise Exception(
-                    "Duplicate key found\nDetail: {}".format(error)
-                )
+                raise Exception("Duplicate key found\nDetail: {}".format(error))
 
         return query
 
     def update(self, id, patches):
+        if isinstance(patches, dict):
+            filteredPatches = clearNone(patches)
+        else:
+            filteredPatches = clearNone(patches.dict())
         with self.session_factory() as session:
             session.query(self.model).filter(
-                self.model.id == id
-            ).update(patches.dict())
-
+                self.model.__dict__[self.identifier] == id
+            ).update(filteredPatches)
             session.commit()
 
     def get(self, id: str) -> T | None:
@@ -59,11 +68,16 @@ class BaseRepo(BaseRepository[T]):
         with self.session_factory() as session:
             query = (
                 session.query(self.model)
-                .filter(self.model.id == id)
+                .filter(self.model.__dict__[self.identifier] == id)
                 .first()
             )
-            session.delete(query)
+
+            if query is None:
+                raise Exception("Element Not found!")
+
+            isDel = session.delete(query)
             session.commit()
+            return isDel
 
 
 # BaseRepo()
